# swagger/swagger.yaml
---
swagger: '2.0'
info:
  version: 1.0.0
  title: Comune di Rimini - Gestione Documentale
  description: |
    # API per l'accesso al servizio di gestione documentale
    Per informazioni contattare Ruggero Ruggeri - 0541/704607 - ruggero.ruggeri@comune.rimini.it
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header  
#host: 10.10.128.79:9988
basePath: /simplesaml/gatewayprotocollo
paths:
  /login:        
    post:
      tags:
        - Autenticazione
      summary: Esegue la richiesta di autenticazione al software di gestione documentale
      description: Per utilizzare le api di protocollazione è necessario ottenere un token di autenticazione attraverso la chiamata a questa API. 
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: Credenziali da inviare per l'autenticazione
          required: false
          schema:
            $ref: "#/definitions/Credenziali"
      responses:
        "200":
          description: OK!
        "401":
          description: Auth error!
        "500":
          description: Invalid input data
  /test:        
   post:
      tags:
        - Autenticazione
      summary: Esegue la richiesta di TEST al software di gestione documentale
      description: Chiamare questa api per verificare il corretto funzionamento dell'autenticazione. Impostare il token di autenticazione, ottenuto con la api /login, premento la i-rossa ed anteponendo al valore il testo 'Bearer'
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      responses:
        "200":
          description: Richiesta corretta vengono ritornati i dati utente
        "401":
          description: Auth error!
        "500":
          description: Invalid input data
      security:
        - Bearer: []  
  /logout:        
   post:
      tags:
        - Autenticazione
      summary: Esegue la richiesta logout dal sistema di gestione documentale
      description: Rende il token di autenticazione inutilizzabile
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      responses:
        "200":
          description: Richiesta corretta vengono ritornati i dati utente
        "401":
          description: Auth error!
        "500":
          description: Invalid input data
      security:
        - Bearer: []          
  /protocollo:
    post:
      tags:
        - Protocollazione
      summary: Esegue la richiesta di protocollazione
      description: La presente api permette la protocollazione "atomica" di uno più documenti. Seguendo le indicazioni del tracciato è possibile inserire i dati del richiedente, il fascicolo e gli allegati
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: I dati JSON utili alla protocollazione
          required: false
          schema:
            $ref: "#/definitions/tracciatoStandard"
      responses:
        "200":
          description: OK!
        "401":
          description: Auth error!
        "500":
          description: Invalid input data
      security:
        - Bearer: []          
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
        description: User Status
  Credenziali:
    type: object
    properties:
      username:
        type: string
        description: nome utente
      password:
        type: string
        description: password
      userCompany:
        type: string
        description: Nome dell'applicativo esterno abilitato all'accesso
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
  tracciatoStandard:
    type: object
    properties:
      richiedente:
        type: string
        $ref: "#/definitions/richiedenteDef"
      datiProtocollo:
        type: object
        properties:
          oggettoDocumento:
            type: string
            default: "Oggetto di prova"
          tipoDocumento:
            type: string            
            default: "01001"
          classificaDocumento:
            type: string
            default: "001.001001"
          origineDocumento:
            type: string
            default: "A"            
          ufficioInternoMittenteDocumento:
            type: string
            default: "404"            
          ufficioInternoDestinatarioDocumento:
            type: string
            default: "404"            
          annoPratica:
            type: string
            default: "2017"
          numeroPratica:
            type: string
            default: "2017-404-0001"
      allegati:
        type: array
        items:
          $ref: "#/definitions/Allegato"
  Tag:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
  Allegato:
    type: object
    properties:
      nomeFile:
        type: string
        default: "METADATI.TXT"
      dimensioneFile:
        type: string
        default: "100"
      mimeType:
        type: string
        default: "text/csv"
      base64:
        type: string
        default: "bG9nX2ZpbGVuYW1lO2xvZy1ERU1PLmxvZw0KbG9nX2xldmVsO0RFQlVHDQpzZWN1cml0eV9jb21w
YW55X2ZpbHRlcjtERU1PDQo="    
      hashSHA256File:
        type: string
        default: "aaaaaa"    
      descrizione:
        type: string
        default: "METADATI"    
  richiedenteDef:
    type: object
    properties:
     nome:
        type: string 
        default: "MARIO"
     cognome:
        type: string
        default: "ROSSI" 
     codiceFiscale:
        type: string 
        default: "RSSMRA00A01H501C"
     dataDiNascita:
       type: string 
       default: "01/01/2000"
     indirizzo:
       type: string 
       default: "VIA ROMA 1"
     citta:
       type: string 
       default: "ROMA"
     mail:
       type: string 
       default: "rossi.mario@roma.com"
     telefonoCellulare:
       type: string 
       default: "333333333333"
     telefonoFisso:
       type: string         
       default: "06 32155531"
     tipoPersona:
       type: string
       default: "FI"         
  Order:
    type: object
    properties:
      id:
        type: integer
        format: int64
      petId:
        type: integer
        format: int64
      quantity:
        type: integer
        format: int32
      shipDate:
        type: string
        format: date-time
      status:
        type: string
        description: Order Status
      complete:
        type: boolean